Fix compiler error about ambiguous int type
conversion from 'std::chrono::duration<long int>::rep' {aka 'long int'} to 'const QJsonValue' is ambiguous

Doesn't happen locally, weirdly enough. On my system and sdk, the chrono
type is a 'long long', which matches the qint64 overload of QJsonValue.

+++ src/token_users.cpp
@@ -283,7 +283,7 @@ int TokenPairingEnableItem::setValue(const QVariant &value)
 		return -1;
 
 	QJsonObject obj;
-	obj["expires_at"] = uptime + pairingDuration.count();
+	obj["expires_at"] = static_cast<int>(uptime + static_cast<int>(pairingDuration.count()));
 	QJsonDocument doc(obj);
 	QString json = doc.toJson(QJsonDocument::Compact);
